+++ cmd/gogoodwe/apilogin/apilogincredentials.go
@@ -21,34 +21,11 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package semsapi
+package apilogin
 
-import (
-	"errors"
-	"net/http"
-	"strings"
-
-	"github.com/AaronSaikovski/gogoodwe/cmd/gogoodwe/types"
-)
-
-// SetHeaders - Set the login headers for the SEMS API login
-func setHeaders(r *http.Request) {
-	r.Header.Add("Content-Type", "application/json")
-	r.Header.Add("Token", "{\"version\":\"v2.1.0\",\"client\":\"ios\",\"language\":\"en\"}")
-}
-
-// CheckUserLoginResponse - check for successful login return value..return a login error
-func checkUserLoginResponse(loginResponse string) error {
-	if strings.Compare(loginResponse, "Successful") != 0 {
-		return errors.New("**API Login Error: " + loginResponse)
-	}
-	return nil
-}
-
-// CheckUserLoginInfo - Check user login struct is valid/not null
-func checkUserLoginInfo(UserLogin *types.LoginCredentials) error {
-	if *UserLogin == (types.LoginCredentials{}) {
-		return errors.New("**Error: User Login details are empty or invalid..**")
-	}
-	return nil
+// ApiLoginCredentials - Struct to hold User login credentials
+type ApiLoginCredentials struct {
+	Account        string `json:"account"`
+	Password       string `json:"pwd"`
+	PowerStationID string `json:"powerstationid"`
 }
