+++ cmd/gogoodwe/apilogin/loginutils.go
@@ -21,12 +21,32 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-package powerstation
+package apilogin
 
-const (
-	// Powerstation API Url
-	PowerStationURL string = "v2/PowerStation/GetMonitorDetailByPowerstationId"
-
-	// Default timeout value
-	HTTPTimeout int = 20
+import (
+	"errors"
+	"net/http"
+	"strings"
 )
+
+// SetHeaders - Set the login headers for the SEMS API login
+func setHeaders(r *http.Request) {
+	r.Header.Add("Content-Type", "application/json")
+	r.Header.Add("Token", "{\"version\":\"v2.1.0\",\"client\":\"ios\",\"language\":\"en\"}")
+}
+
+// CheckUserLoginResponse - check for successful login return value..return a login error
+func checkUserLoginResponse(loginResponse string) error {
+	if strings.Compare(loginResponse, "Successful") != 0 {
+		return errors.New("**API Login Error: " + loginResponse)
+	}
+	return nil
+}
+
+// CheckUserLoginInfo - Check user login struct is valid/not null
+func checkUserLoginInfo(userLogin *ApiLoginCredentials) error {
+	if *userLogin == (ApiLoginCredentials{}) {
+		return errors.New("**Error: User Login details are empty or invalid..**")
+	}
+	return nil
+}
