+++ cmd/gogoodwe/monitordata/interfaces.go
@@ -21,30 +21,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
+package monitordata
 
-
-/*
-# Name: DailySummaryData - Struct to hold daily summary data
-*/
-package types
-
-type DailySummaryData struct {
-	Language string `json:"language"`
-	HasError bool   `json:"hasError"`
-	Msg      string `json:"msg"`
-	Code     string `json:"code"`
-	Data     struct {
-		Kpi struct {
-			MonthGeneration float64 `json:"month_generation"`
-			Power           float64 `json:"power"`
-			TotalPower      float64 `json:"total_power"`
-			DayIncome       float64 `json:"day_income"`
-			TotalIncome     float64 `json:"total_income"`
-			Currency        string  `json:"currency"`
-		} `json:"kpi"`
-		Inverter []struct {
-			TotalGeneration string `json:"total_generation"`
-			DailyGeneration string `json:"daily_generation"`
-		} `json:"inverter"`
-	} `json:"data"`
+// Constraints for functions that return data from the API via marshalled structs
+type ISemsDataConstraint interface {
+	InverterData | DailySummaryData
 }
